ref: extract comment logic out
