+++ package.json
@@ -33,6 +33,7 @@
     ]
   },
   "devDependencies": {
+    "@actions/github": "^4.0.0",
     "@auto-it/conventional-commits": "^10.18.4",
     "@auto-it/git-tag": "^10.18.4",
     "@auto-it/released": "^10.18.4",
